Fix #62 - GitHub user migration retry
Attempt at creating a user migration after 5 minutes, if it fails.
This is of course quick and dirty to validate if this will work.

The interesting thing is the migration does work on the GitHub side,
since i get an email saying the migration is ready and I can download
it.