Always use token for cloning via HTTPS
On GitLab for some reason the `repo.Private` always returns false even if the repo is actually private, which means this code is never executed and the `git` client asks for a username and password to clone. 

It would make sense to use the token for cloning any type of repo, as it would prevent the need for asking these details.

Tested on my local build with gitlab.com and it works as expected.